Region can show alarm and safe zone. Add chernihiv region.
On creation Region is showing the region image in white. It is not saving the bitmap in RAM, only in flash. On alarm, region bitmap is copied to RAM and color is changed to RED, than showed to the display and delited from RAM.
Every region header file is containing corresponding region instance. Not sure if it is the best way, but it will be so for now.